Little bit of JS added to query Parse database for public books and create corresponding objects. DOM elements created in same <article> used for hard coded HTML. Not appended yet, but logged in console if you want to check it out.
We'll have to figure out something for Next and Previous book pages. Only 9 can be displayed on the home page in the present state, so we can either make multiple pages of 3X3 grids or try something else. Once we agree on this, I can append the book elements to the DOM and you'll see them live.

Also also, at that time it's probably best for an intern to populate the cover images in the database. Currently, we only have one good cover, and I'm not even sure if its the one we want to go with. I can give the Parse database credentials and step someone through it on the phone. It's not too bad.